+++ react/components/WordpressPost.tsx
@@ -120,7 +120,7 @@ const WordpressPost: FunctionComponent = props => {
 
     const titleHtml = insane(title.rendered, sanitizerConfig)
     const captionHtml =
-      featured_media?.caption?.rendered ??
+      featured_media?.caption?.rendered &&
       insane(featured_media.caption.rendered, sanitizerConfigStripAll)
     const bodyHtml = insane(content.rendered, sanitizerConfig)